requirements.txt
Django==5.2.3
djangorestframework==3.16.0
pillow==11.2.1
django-filter==25.1
